organize fileproperties functionality into a sub module
